Add dummy.txt to Abc.Directory.Standalone
A NuGet limitation (http://nuget.codeplex.com/workitem/595) prevents
the package from being installed if it has an empty lib folder while
depending on packages that don't. Adding a dummy.txt file to lib allows
the package to be installed, without actually adding the file to the
target project references since it is a .txt file.