Use Peer timestamp to remove dynamic subscriptions
When registering, the local time was used to remove any previously
persisted dynamic subscriptions. Since the peer timestamp was used to
insert subscriptions later, if the directory was in the future, it
caused the delete to prevent the insertion of future subscriptions.
(When using Cassandra as a persistence, because it resolves conflicts
using timestamps).