Add WasPersisted to TransportMessage
Adding the WasPersisted field to the TransportMessage allows the
receiver to know if the message was sent to the persistence and needs
to be acknowledged.
Before that the receiver had to rely on its knowledge of the
persistence setting of the message, which could be different between
sender and receiver and cause unneeded acks or unacked messages.