Add MessageId to clustering key
The previous clustering key only included the time component of the
MessageId, supposing that timestamps were unique. This was safe to
assume for a single Peer, since timestamps are guarantied to be unique
for a given Peer by Zebus, but it presented a risk of message overwrite
if two Peers sent a message to the same destination at the exact same
time.