Replace owned thread in PeriodicActionHostInitializer by a timer
It's basically the same things as before, the only notable change is that the periodic action code can now be executed by different threads (from the threadpool). But at least the periodic action invoker makes sure there is no multiple executions at the same if the periodic action takes to much time.
I know there is a better way to handle the concurrent executions problem by using the timer without a period but only with a due time but I found this solution a bit ugly and complicated (the code would be a bit clunky IMO). The issue with the implem I made is that if the periodic action is not finish within the given period we will have to wait for another full period to retry ... which is a good thing I guess.