Allow SubscriptionSnapshotGenerator to skip sending a snapshot.
It's already possible for SubscriptionMultipleSnapshotGenerator to skip sending a snapshot
(by returning an empty list). Being able to do the same for a single snapshot seems reasonable.

One use case for this is if the snapshot message is not always enabled (for example if it is
guarded behind a launch/feature flag).