Remove ZmqTransport.SocketOptions property
ZmqSocketOptions is mutable but it is not clear when it is possible to
change it or when it is too late. It is much cleaner to keep this member
private.

The property was public but it was not part of ITransport. It is highly
unlikely that this change would break code somewhere.