Directory.Cassandra: Update schema of `Peers` and `DynamicSubscriptions` tables to add multiple partitions.
Previous schema was using a "fake" boolean key as a unique partition. This meant that all the peers and their
subscriptions were stored in a single partition.

Every time we needed to remove a peer and its associated subscriptions, we had to tombstone a cell from
a single partition, which was triggering a warning from Cassandra when scanning peers with too many
tombstones (cassandra.yaml `tombstone_warn_threshold`)

We now partition the data by PeerId which makes it possible to tombstone a whole partition instead of
individual cells in a row