Use ReadOnlyMemory<byte> instead of Stream for TransportMessage content
The goal of Stream was to encapsulate the underlying buffer and allow
future optimizations, for example by using RecyclableMemoryStream
to reduce buffer allocations.

Yet, these optimizations were never implemented. Also, streams are
stateful and non-thread-safe, which makes passing TransportMessage
among threads quite dangerous. So, a clean implementation of the stream
design would imply to properly manage TransportMessage ownership,
to copy them when required and to dispose the underlying stream.

A concurrency bug was identified in PersistMessageCommandHandler,
because a TransportMessage was forwarded to multiple threads. Instead of
fixing this very specific bug, it was decided to make TransportMessage
immutable by using ReadOnlyMemory instead of Stream, thus removing
the possibility of the currency bug to happen.

This is a breaking change that should only impact low level code.