Support auto-subscribe routable message
The default subscription mode of routable message is Manual.
This behavior has two issues: it breaks the principle of least
surprise and it makes turning existing message routable a dangerous
operation.

Changing this behavior now is not possible, but it would be nice to
configure new routable messages with the Auto subscription mode.

This change allows to configure the default subscription mode of
routable messages.